worker-sim/pyproject.toml
[project]
name = "worker-sim"
version = "0.1.0"
description = "Simulated device worker for EdgeOrchestra"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.68.0",
    "protobuf>=5.29.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
    "orchestrator",
]

[tool.uv.sources]
orchestrator = { workspace = true }

[project.scripts]
worker-sim = "worker_sim.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/worker_sim"]
